Re-add model loading and clean up code

+++ src/scene_manager.cpp
@@ -25,7 +25,7 @@ namespace engine {
 	void SceneManager::updateActiveScene(float ts)
 	{
 		if (m_activeSceneIndex >= 0) {
-			assert(m_activeSceneIndex < m_scenes.size());
+			assert((size_t)m_activeSceneIndex < m_scenes.size());
 			m_scenes[m_activeSceneIndex]->update(ts);
 		}
 	}
